+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Drop support for Java 7 and Clojure(Script) 1.7.
 * Rename `cider-eval-defun-to-point` to `cider-eval-defun-up-to-point`.
 * Add support for printing to the current buffer to `cider-eval-defun-up-to-point`.
+* Remove `cider-ping` command.
 
 ## 0.17.0 (2018-05-07)